Adding redux-toolkit and a store. Starting with lesson reducer, and for LessonsPage.
The names under createAsyncThunk don't mean something to us, they're the names of the actions, so redux toolkit handles them internally.

The reducers use the api that Tomer wrote, and handles the response of the promise.

important in dispatch to do useDispatch<AppDispatch>(), so it knows the types of what's defined inside the dispatch and it's not casting as any.