Fix canHaveTimer so now it can be shown whenever the quiz finished loading.
When handling submittion (of normal attempt, not on each answer), save the newly-created attempt id in the state.

Returned loading to false in initial and setIsLocked to submittion, those were mistakes. Without setIsLocked it won't lock when it's a normal saving, so I want it to lock if it succeeds saving, but I want it to loc no matter what if the timer runs out, even if it fails saving.