Fixed creating an empty attempt when it's onSubmit quiz mode.
Fixed retry, changing to no attempt so it creates an empty new one.

Creating QuizTimer, shared both in the quiz itself and QuizItem. Made most of its properties optional so in QuizItem I just display the time and change the time with the interval.

Instead of quizResult and areAllQuestionsSubmitted, having a boolean prop. Put setIsLocked true inside onTimeUp, so no need to pass it.

Adding expiryTime to  attempt.