Only have the timer if currentAttempt exists, otherwise it initialized the timer interval with undefined initial time.
In QuizItem, only show the timer if the attempt is incomplete.

Creating an empty attempt also for normal submission. Created route updateAttemptWithAnswers, which gets an attempt id and swithces all its answers with the ones I'm giving (didn't want to complicate add answer logic).
Initial time in timer is now based on that empty attempt time.