Fixed bug of quiz not being kept in refresh, because redux doesn't save it in refreshes.
So now if I have a quizId correctly but the quiz data from the redux is empty, try re-fetching it.

Reverted back the InvalidNavigationGuard. As Tomer said, it could lead to an inconsistent order/number of hooks, break the rules of hooks and break the entire app, not worth the risk. I may return it later, for now no need.