Fix NotificationPage having glithces when fetching notifications, by knowing when it's not the first loading.
Fixed bug I just created, in which marking a notification as read closed the computer notification and as a result it deleted the app notification. Now I remove the onclose side effects before closing the computer notification.

Reduced the default notification time for a nice effect of the computer getting the notifications quickly.